Move MetaInformation to model package
